cheri: Add missing pointer cast annotations in mm
Add missing annotations for pointer and uintptr_t related
casts in mm/.

+++ mm/debug.c
@@ -131,7 +131,7 @@ static void __dump_page(const struct page *page)
 {
 	struct folio *foliop, folio;
 	struct page precise;
-	unsigned long head;
+	uintptr_t head;
 	unsigned long pfn = page_to_pfn(page);
 	unsigned long idx, nr_pages = 1;
 	int loops = 5;
